feat: move aggregate functions to new file
